+++ web/entrypoints/u.main.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { UsernamePage }  from "../components/UsernamePage"
-import { UArgs } from "./shared";
+import { UArgs } from "../shared/types";
 
 declare var uargs: UArgs;
