+++ web/shared/types.ts
@@ -1,0 +1,36 @@
+export interface Account {
+  username: string
+  apiKey: string
+  private: boolean
+}
+
+export interface BootstrapArgs {
+  host: string
+  email: string
+  loginURL: string
+  logoutURL: string
+  account: Account
+}
+
+export interface UArgs {
+  host: string
+  artworkBaseURL: string
+  profileUsername: string
+  logoutURL: string
+  account: Account
+  self: boolean
+}
+
+export interface Song {
+  albumTitle: string
+  artistName: string
+  title: string
+  totalTime: number
+  year: number
+  lastPlayed: number
+  artworkHash: string
+  trackViewURL: string
+  loved: boolean
+
+  ident: string
+}
