+++ web/shared/util.ts
@@ -1,42 +1,5 @@
 import * as base64js from "base64-js";
 
-export interface Account {
-  username: string
-  apiKey: string
-  private: boolean
-}
-
-export interface BootstrapArgs {
-  host: string
-  email: string
-  loginURL: string
-  logoutURL: string
-  account: Account
-}
-
-export interface UArgs {
-  host: string
-  artworkBaseURL: string
-  profileUsername: string
-  logoutURL: string
-  account: Account
-  self: boolean
-}
-
-export interface Song {
-  albumTitle: string
-  artistName: string
-  title: string
-  totalTime: number
-  year: number
-  lastPlayed: number
-  artworkHash: string
-  trackViewURL: string
-  loved: boolean
-
-  ident: string
-}
-
 export function hasPrefix(s: string, prefix: string): boolean {
   return s.length >= prefix.length && s.slice(0, prefix.length) == prefix
 }
